fix delete functionality

+++ src/components/DeleteProject.tsx
@@ -1,11 +1,22 @@
 "use client";
 
-import React, { FC } from "react";
+import React, { FC, useState } from "react";
 import { Button } from "@/components/ui/button";
 import { useMutation } from "@tanstack/react-query";
 import axios from "axios";
 import { useToast } from "@/components/ui/use-toast";
 import { useRouter } from "next/navigation";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 interface DeleteProjectProps {
   projectId: string | undefined;
@@ -18,7 +29,7 @@ const DeleteProject: FC<DeleteProjectProps> = ({ projectId }) => {
   const { mutate: deleteProject, isLoading } = useMutation({
     mutationFn: async (id: string | undefined) => {
       if (id) {
-        const { data } = await axios.delete("/api/" + id);
+        const { data } = await axios.delete("/api/projects/" + id);
         return data;
       }
     },
@@ -41,13 +52,32 @@ const DeleteProject: FC<DeleteProjectProps> = ({ projectId }) => {
   });
 
   return (
-    <Button
-      variant="destructive"
-      isLoading={isLoading}
-      onClick={() => deleteProject(projectId)}
-    >
-      delete
-    </Button>
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <Button variant="destructive">delete</Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. This will permanently delete this
+            project and remove its data from our servers.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <AlertDialogAction asChild>
+            <Button
+              variant="destructive"
+              isLoading={isLoading}
+              onClick={() => deleteProject(projectId)}
+            >
+              delete
+            </Button>
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 };
